pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "myclaw"
version = "0.1.0"
description = "Autonomous AI agent powered by Claude Code -p (Max subscription)"
requires-python = ">=3.11"
dependencies = [
    "hyperliquid-python-sdk>=0.4.0",
    "pyyaml>=6.0",
    "jsonschema>=4.20",
    "requests>=2.31",
    "python-dotenv>=1.0",
]


[project.optional-dependencies]
gateway = [
    "python-telegram-bot>=21.0",
    "aiohttp>=3.9",
]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
